obs-ffmpeg: Use output_video_info instead of obs_video_info in amf_create_encoder()
The frame rate used to initialize an AMF encoder should be aligned
with the derived frame rate in output_video_info instead of the global
obs_video_info structure. With this change IDR's can be aligned when
multiple renditions are being encoded.

Using video_output_info members for the format, colorspace, and range
parameters in addition to the frame rate provides a single source for
this information and obs_video_info is no longer needed.