libobs: Insert captions on all video tracks
We will now keep a per track list of caption data. When caption
insertion is triggered we will add that to the list for each
actively encoding video track. When doing interleaved send we will
pull the data from the caption data for the corresponding video.
This ensures that captions get copied to all video tracks.