libobs: Add encoded_frames counter
For metrics reporting it is important to know the
number of frames successfully encoded by the
obs_encoder_t instances, particularly for renditions
where the encoded frame rate differs from the canvas
frame rate. The encoded_frames counter is introduced
to measure and report this metric via SEI in the
encoded rendition metrics message.