libobs: Remove the BPM frame tracker mutex
Tracking frame-level timestamp information for the
purposes of generating BPM timestamps requires a
queue (implemented as an OBS DARRAY). Initially a
mutex was used to protect concurrent access to the
queue, however it was determined to not be necessary
once the thread model and access patterns were understood,
and the removal of the mutex was tested.