obs-ffmpeg: Set the codec level properly for AMF encoders
The default "level" setting was being used for each codec
(AVC, HEVC, AV1) supported by AMF. For example, all HEVC
encoders were using level 6.2 and this caused some playback
devices to reject the bitstream for decode because the device
reported a maximum decode level lower than 6.2.

Add functionality to determine the best match for the codec
level instead of relying on the defaults.