obs-ffmpeg: Make VAAPI profile handling string-based
VAAPI support was using an integer for the codec profiles instead
of strings. This made it incompatible with enhanced broadcasting
where a string is sent for the profile. The logic for handling
codec profiles is now using strings, and an appropriate conversion
from string to the required integer is also implemented.

FFmpeg has deprecated the FF_PROFILE_* definitions in favor of
AV_PROFILE_* definitions, which are now being used with this commit.