CI: Add Twitch hash and client_id to flatpak builds
The flatpak build for TEB on Linux did not have the
"Connect Account" option enabled in the Auto Config
Wizard. This option is enabled only when the TWITCH_HASH
and TWITCH_CLIENT_ID are present in the environment during
build. Enable the option in the build-project workflow for
flatpak builds.