UI: Rework Linux GPU association logic
With enhanced broadcasting on Linux the logic
used to associate the GPU OBS is using with the
libpci-scanned information was inadequate in some cases.
Rework the solution to use a string-matching algorithm
that tokenizes the GPU identification string then compares
it to the strings returned by libpci. This solves the
issue of associating the in-use GPU with the wrong adapter
in the list sent to the server in the JSON request.