UI: Fix TEB Linux GPU card probing
To detect DRM (Direct Rendering Manager) cards in Linux, the
implementation is scanning `/sys/class/drm`. Originally the
code was looking for `card<n>` (where n is 0,1,2,...), but this
turned out to be risky because there is no guarantee about the
index starting at 0. Rework the code to scan through every entry
in the `/sys/class/drm/` directory and use the `renderD<nnn>`
entries, where nnn usually starts at 128 but the value is now
being ignored to avoid any further indexing issues.