build-aux: Fix incorrect HEVC vertical resolution on AMD GPUs
Please refer to [1], [2], and [3] for context. The patches are taken from [2].

AMD GPUs use a different surface alignment compared to other GPUs for encoding,
which required an upstream fix in Mesa, libva, FFmpeg, and the VCN firmware.

This commit adds the FFmpeg patches to the flatpak build of OBS which is currently
using FFmpeg version n7.1, however the patches were applied upstream after that
release. libva API version 1.21 or higher is also required, which is satisfied by
using KDE runtime 6.8.

The incorrect 1920x1088 HEVC output resolution (instead of 1920x1080) resulted in
streaming disconnects for Twitch Enhanced Broadcsating Linux users. Local file
recording when using AMD VAAPI HEVC also outputs 1088 instead of 1080. This commit
fixes both cases.

Note that the fix is applicable to flatpak builds only. Standalone and Debian
package builds depend on the host library versions and will still have the problem
until a version of FFmpeg with the fix is used.

[1]: https://patchwork.ffmpeg.org/project/ffmpeg/list/?series=11222
[2]: https://patchwork.ffmpeg.org/project/ffmpeg/list/?series=13168
[3]: https://gitlab.freedesktop.org/mesa/mesa/-/issues/10985