build-aux: Use more recent patch for AMD HEVC VAAPI fix
Use the patch series in [1] to fix the AMD HEVC vertical resolution issue
where 1920x1088 is output instead of 1920x1080. Originally I had used
an older upstream patch that will not be merged in FFmpeg.

This new patch series did not apply cleanly as-is. Minor edits were
made to the patch and a combined sinlge diff file is being used for our
beta build until it is merged upstream.

[1]: https://patchwork.ffmpeg.org/project/ffmpeg/list/?series=13168