CI: Disble flatpak builds
TEB-on-Linux support was merged upstream as of OBS 31.0,
so we don't need to build and distribute flatpaks in the
TEB beta any longer. Also, the flatpak builds using the
GitHub runners for our service tier were often failing
for different reasons, the main one being lack of disk
space in the runner.